+++ src/endpoints/HighestMarginsWrapper.ts
@@ -1,8 +1,0 @@
-import APIBaseWrapper from './APIBaseWrapper';
-import { ItemListParams, ItemListPromise } from '../types';
-
-export default class HighestMarginsWrapper extends APIBaseWrapper {
-    getItems(opts: ItemListParams = { filters: false }): ItemListPromise {
-        return this._wrapGet(this.parseOptions('highest-margins', opts));
-    }
-}