+++ src/endpoints/index.ts
@@ -1,23 +1,16 @@
 import APIBaseWrapper from './APIBaseWrapper';
 import APIUptimeWrapper from './APIUptimeWrapper';
-import BarrowsRepairWrapper from './BarrowsRepairWrapper';
-import CombinationItemsWrapper from './CombinationItemsWrapper';
 import CraftingSmithingWrapper from './CraftingSmithingWrapper';
 import DashboardWrapper from './DashboardWrapper';
-import DecantPotionsWrapper from './DecantPotionsWrapper';
 import FavouriteItemsWrapper from './FavouriteItemsWrapper';
+import FlipFinderWrapper from './FlipFinderWrapper';
 import FletchingWrapper from './FletchingWrapper';
-import GELimitsWrapper from './GELimitsWrapper';
 import GraphWrapper from './GraphWrapper';
 import HerbloreWrapper from './HerbloreWrapper';
-import HighVolumeWrapper from './HighVolumeWrapper';
-import HighestMarginsWrapper from './HighestMarginsWrapper';
-import ItemSetsWrapper from './ItemSetsWrapper';
 import ItemsWrapper from './ItemsWrapper';
 import LeaderboardWrapper from './LeaderboardWrapper';
 import MagicWrapper from './MagicWrapper';
 import MarketWatchWrapper from './MarketWatchWrapper';
-import NewItemsWrapper from './NewItemsWrapper';
 import NotificationsWrapper from './NotificationsWrapper';
 import PriceAlertWrapper from './PriceAlertWrapper';
 import ProfitTrackerWrapper from './ProfitTrackerWrapper';
@@ -31,24 +24,17 @@ import UserWrapper from './UserWrapper';
 export {
     APIBaseWrapper,
     APIUptimeWrapper,
-    BarrowsRepairWrapper,
-    CombinationItemsWrapper,
     CraftingSmithingWrapper,
     DashboardWrapper,
-    DecantPotionsWrapper,
     FavouriteItemsWrapper,
+    FlipFinderWrapper,
     FletchingWrapper,
-    GELimitsWrapper,
     GraphWrapper,
     HerbloreWrapper,
-    HighVolumeWrapper,
-    HighestMarginsWrapper,
-    ItemSetsWrapper,
     ItemsWrapper,
     LeaderboardWrapper,
     MagicWrapper,
     MarketWatchWrapper,
-    NewItemsWrapper,
     NotificationsWrapper,
     PriceAlertWrapper,
     ProfitTrackerWrapper,